Changed the hpx_addr_t representation to what we need for the new PGAS.
The block size has to be explicitly passed to hpx_addr_add now, and hence
this patch adjusts all of its occurrences accordingly.

+++ apps/libPXGL/include/dimacs.h
@@ -16,8 +16,9 @@
 #include "hpx/hpx.h"
 #include "adjacency_list.h"
 
+
 // Compute checksum given an adjacency list
 extern hpx_action_t dimacs_checksum;
-extern int dimacs_checksum_action(const size_t *const g);
+extern int dimacs_checksum_action(const uint64_t *const g);
 
 #endif // PXGL_DIMACS_H
